+++ R/ctx_node_vlmc.R
@@ -53,6 +53,12 @@ cutoff.ctx_node <- function(model, scale = c("quantile", "native"), raw = FALSE,
 #'
 #' @param model T `ctx_node` object as returned by [find_sequence()].
 #' @inherit metrics
+#' @examples
+#' pc <- powerconsumption[powerconsumption$week == 5, ]
+#' dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.25, 0.5, 0.75, 1))))
+#' model <- vlmc(dts)
+#' model_ctxs <- contexts(model)
+#' metrics(model_ctxs[[4]])
 #' @export
 metrics.ctx_node <- function(model, ...) {
   c_probs <- matrix(probs(model), nrow = 1)